+++ Realm/Swift/RLMSupport.swift
@@ -20,14 +20,15 @@ import Realm
 
 #if swift(>=3.0)
     extension RLMRealm {
-        public class func schemaVersion(at url: URL, usingEncryptionKey key: Data? = nil) throws -> UInt64 {
-            var error: NSError?
-            let version = __schemaVersion(at: url, encryptionKey: key, error: &error)
-            guard version != RLMNotVersioned else {
-                throw error!
-            }
-            return version
-        }
+        // TODO: Figure out why this causes the Swift 3 compiler to segfault.
+//        public class func schemaVersion(at url: URL, usingEncryptionKey key: Data? = nil) throws -> UInt64 {
+//            var error: NSError?
+//            let version = __schemaVersion(at: url, encryptionKey: key, error: &error)
+//            guard version != RLMNotVersioned else {
+//                throw error!
+//            }
+//            return version
+//        }
     }
 
     extension RLMObject {
@@ -55,7 +56,7 @@ import Realm
         }
     }
 
-    // TODO: Figure out why this causes the Swift 3 compiler to segfault.
+    // SR-2348: A bug in Objective-C generics currently make this impossible w/o an error or compiler crash.
 //    extension RLMArray: Sequence  {
 //        // Support Sequence-style enumeration
 //        public func makeIterator() -> RLMIterator {